+++ fe/fe-core/src/main/java/org/apache/doris/planner/JdbcScanNode.java
@@ -170,7 +170,7 @@ public class JdbcScanNode extends ScanNode {
         numNodes = numNodes <= 0 ? 1 : numNodes;
 
         StatsRecursiveDerive.getStatsRecursiveDerive().statsRecursiveDerive(this);
-        cardinality = statsDeriveResult.getRowCount();
+        cardinality = (long) statsDeriveResult.getRowCount();
     }
 
     @Override
